project/requirements.txt
django==2.1.4
djangorestframework==3.9.0
pillow==5.3.0
django-rest-swagger==2.2.0
djangorestframework-simplejwt==3.3
django-cors-headers==2.4.0
gunicorn==19.9.0
psycopg2==2.7.3.1
flake8==3.7.8
ipython==7.7.0
sentry-sdk==0.10.2
django-mailgun==0.9.1
python-decouple==3.1
django-anymail==7.0.0